% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAFSIL_utils.R
\name{fast.pca}
\alias{fast.pca}
\title{PCA built on fast.rsvd}
\usage{
fast.pca(X)
}
\arguments{
\item{X}{matrix}
}
\value{
projection of X on principal top principal components
}
\description{
PCA built on fast.rsvd
}
