% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAFSIL_utils.R
\name{fast.rsvd}
\alias{fast.rsvd}
\title{RANDOM PROJECTION SVD}
\usage{
fast.rsvd(A, K)
}
\arguments{
\item{A}{matrix}

\item{K}{rank}
}
\value{
list with components U S and V
}
\description{
RANDOM PROJECTION SVD
}
