% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAFSIL_utils.R
\name{plotSIM}
\alias{plotSIM}
\title{Plot a (dis)similarity matrix using the heatmap function}
\usage{
plotSIM(x, labels, col = "qnt")
}
\arguments{
\item{x}{matrix (Dis)Similarities}

\item{labels}{integer Labels for classes/categories}

\item{col}{character "qnt" for quantile-based color breaks, "lin" for linear color breaks}
}
\value{
list The list of components heatmap returns.
}
\description{
Plot a (dis)similarity matrix using the heatmap function
}
