% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAFSIL_utils.R
\name{plotTSNE}
\alias{plotTSNE}
\title{Plot a tSNE plot of a (dis)similarity matrix}
\usage{
plotTSNE(x, labels, seed = 3749829, ...)
}
\arguments{
\item{x}{matrix (Dis)Similarities}

\item{labels}{integer Labels for classes/categories}

\item{seed}{integer seed for RNG (->tSNE)}

\item{...}{further arguments to Rtsne}
}
\description{
Plot a tSNE plot of a (dis)similarity matrix
}
