% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAFSIL.R
\name{RAFSIL}
\alias{RAFSIL}
\title{RAFSIL}
\usage{
RAFSIL(data, NumC = NULL, nrep = 50, method = "RAFSIL1",
  gene_filter = TRUE)
}
\arguments{
\item{data}{an (n x m) data matrix of gene expression measurements of single cells}

\item{NumC}{number of clusters to be extracted over data}

\item{nrep}{integer for RAFSIL1 the number of times similarities are learned}

\item{method}{character "RAFISIL1" or "RAFSIL2"}

\item{gene_filter}{logical should gene filtering be performed?}
}
\value{
list of 3 elements describing the results:
	SF:  Spearman Feature Space,
 	D:   RAFSIL dissimilarity matrix,
   lab: inferred cluster labels
 	hc:  hierarchical clustering solution
}
\description{
perform the RAFSIL dissimilarity learning and clustering algorithm
}
