% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafsil_FE.R
\name{rafsilFE}
\alias{rafsilFE}
\title{FEATURE ENGINEERING FOR RAFSIL}
\usage{
rafsilFE(texpr, gene_filter = TRUE, frq = 0.06)
}
\arguments{
\item{texpr}{matrix Expression values, cells x genes}

\item{gene_filter}{logical shold gene filtering be performed?}

\item{frq}{numeric Frequency filtering fraction}
}
\value{
list with two components: feature matrix F and the result of gene clustering cl.gene
}
\description{
FEATURE ENGINEERING FOR RAFSIL
}
